==> nb03_Imputacao_Vendas.ipynb <==
# Coleta, Tratamento e Unificação das bases de dados relativas as Usinas Sucroalcooleiras

# Setup Inicial

Carregando linhas com configurações iniciais já escritas em notebook base

No bloco de comando "%run" não pode ter nenhum outra linha de comando ou comentário

%run nb00_Setup_Usinas{}

# Leitura dos Dados 

usinas   = pd.read_parquet(abfss_path_enriched + 'usinas/usi02_tancagens_impt_usinas.parquet',
                            storage_options = {'linked_service' : linked_service_enriched})

prop  = pd.read_parquet(abfss_path_enriched + 'volumetria/vol01_prop_consfinal_vegetativo.parquet',
                        storage_options = {'linked_service' : linked_service_enriched})

cf  = pd.read_parquet(abfss_path_enriched + 'volumetria/vol01_vendas_vibra_mes.parquet',
                        storage_options = {'linked_service' : linked_service_enriched})

diesel_mes = pd.read_parquet(abfss_path_enriched + 'anp/anp_diesel_mes.parquet',
                                storage_options = {'linked_service' : linked_service_enriched})

hist = pd.read_parquet(abfss_path_enriched + 'histdiesel/hist_diesel.parquet',
                        storage_options = {'linked_service' : linked_service_enriched})


vol_usinas = pd.DataFrame(diesel_mes.loc['2016':]['consumidor final']*prop['USINA'][:-1])

# Ajuste do Histórico de Vendas da VIBRA

# col_names = pd.read_csv(arq_vol, nrows=0).columns                               # coletando informações de cabeçalho
# col_types = {col: str for col in col_names if col not in {}}                    # todas as colunas como str





# #hist.drop(['Unnamed: 0','index'],axis=1,inplace=True) # var não necessária
# hist['VEDI_VOL'] = hist['VEDI_VOL'].astype(float)
# hist['VEDI_DAT_VEN'] = pd.to_datetime(hist['VEDI_DAT_VEN'])

#organizando o Coreto!
hist = hist[['CLIE_NUM_CPF_CNPJ', 'CLIE_NOM_CLIENTE','VEDI_COD_CLI', 'VEDI_VOL', 'VEDI_DAT_VEN', 'ProdutoD', 'AECO']]

hist.rename(columns={'CLIE_NUM_CPF_CNPJ': 'CNPJ'}, errors='raise', inplace=True)


# hist['CNPJ'] = hist['CNPJ'].astype(str).apply(lambda x: x.zfill(14))            # Transforma CPNJ em string
# hist = hist[~hist['CNPJ'].str.startswith('000')]                                # removendo possiveis CPFs

hist = hist.merge(usinas[['CNPJ','UF']], how='left', on ='CNPJ')                # trazendo infos de Usinas Gov para filtro

hist = hist[hist['VEDI_VOL'] > 0].sort_values('VEDI_DAT_VEN')                   # Eliminando Vendas negativas ou zeradas

hist = hist[~hist['UF'].isna()].reset_index(drop=True)                          # Filtrando CNPJ que estão na base Usinas - independente da AECO

### Agrupando vendas por mês

hist['ANOMES'] = hist['VEDI_DAT_VEN'].dt.strftime('%Y-%m')

hist_m = hist.groupby(['CNPJ','ANOMES'])['VEDI_VOL'].agg(Vol_Total = 'sum', Freq='count', Vol_Medio='mean').reset_index()
hist_m['data'] = pd.to_datetime(hist_m['ANOMES'], format='%Y-%m-%d')
hist_m = hist_m.set_index('data').sort_index()

### Incluindo vendas 0 nos meses que não houveram vendas

hist_e = pd.DataFrame()
periodo = pd.date_range(start='2016-01-01', end=dt.today(), freq='MS')

for x in periodo.year.unique():
    per = [a for a in periodo if a.year==x]
    aux = (hist_m.loc[str(x)].groupby('CNPJ')
                 .apply(lambda x: x.reindex(per).fillna(0))  # preenche os meses sem vendas com 0
                 .rename_axis(('CNPJ','Data'))
                 .drop(['CNPJ','ANOMES'], axis=1)
                 .reset_index())
    hist_e = pd.concat([hist_e, aux], axis=0, ignore_index=True)
    
hist_e = hist_e.merge(hist[~hist['CNPJ'].duplicated()][['CNPJ','AECO']], how='left', on='CNPJ')   # resgatando a AECO

# Filtrando histórico de Clientes Vibra Contidos na Base de Usinas do GOV

df = usinas[['CNPJ', 'MUNICIPIO','UF', 'CAP_HIDRATADO','CAP_ANIDRO','TANCAGEM_ETANOL','TANCAGEM_DIESEL', 'SITUACAO_ANP']].merge(
       hist_e, how='left', on ='CNPJ')

df = df[~df['Data'].isna()]                                                                       # somente os clientes vibra dentro do MAPA+ANP
df['ANOMES'] = df['Data'].dt.strftime('%Y-%m')
df['ANO'] = df['Data'].dt.strftime('%Y')

# Imputação das Vendas

### Imputação Mensal

# Imputação ANO MES

cols = ['CNPJ','UF','REGIAO','Data', 'ANOMES', 'ANO', 'CAP_HIDRATADO','CAP_ANIDRO','TANCAGEM_ETANOL','TANCAGEM_DIESEL', 'Vol_Total','Freq','Vol_Medio']
var  = ['CAP_HIDRATADO','CAP_ANIDRO','TANCAGEM_ETANOL','TANCAGEM_DIESEL','Vol_Total','Freq','Vol_Medio']

res_m = imp_anual(df, freq='ANOMES', l=['CNPJ','Data','ANOMES', 'ANO'], cols=cols, var=var, ind=False)

res_m['ANO']= [ x[:4] for x in res_m['ANOMES']]
res_m['IMP'] = [1 if pd.isnull(x) else 0 for x in res_m['Data']]
res_m['Data'] = pd.to_datetime(res_m['ANOMES'], format='%Y-%m-%d')

reg = {'N':'N-NE','NE':'N-NE','S':'CO-S-SE','CO':'CO-S-SE', 'SE':'CO-S-SE'}

res_m['REGIAO2'] = [reg[x] for x in res_m['REGIAO']]
res_m['REGIAO3'] = np.where(res_m['UF']=='SP', res_m['UF'], res_m['REGIAO2'])

### Imputação Anual

df_ga = df.groupby(['CNPJ','ANO']).sum().reset_index()

cols = ['CNPJ','UF','REGIAO','ANO', 'CAP_HIDRATADO','CAP_ANIDRO','TANCAGEM_ETANOL','TANCAGEM_DIESEL', 'Vol_Total','Freq','Vol_Medio']
var  = ['CAP_HIDRATADO','CAP_ANIDRO','TANCAGEM_ETANOL','TANCAGEM_DIESEL','Vol_Total','Freq','Vol_Medio']

res_a = imp_anual(df_ga, 'ANO',['CNPJ','ANO'],cols, var, ind=False)

res_a['REGIAO2'] = [reg[x] for x in res_a['REGIAO']]
res_a['REGIAO3'] = np.where(res_a['UF']=='SP', res_a['UF'], res_a['REGIAO2'])

# Salvando dados

res_m.to_parquet(abfss_path_enriched + 'usinas/usi03_vendas_impt_mes.parquet', 
                     storage_options = {'linked_service':linked_service_enriched})
res_a.to_parquet(abfss_path_enriched + 'usinas/usi03_vendas_impt_ano.parquet', 
                     storage_options = {'linked_service':linked_service_enriched})

def agrup_graf(df, geo='UF'):
    aux = df.set_index('Data')
    aux = aux.groupby([geo])['Vol_Total'].resample('MS').sum()
    aux = aux.unstack(level=0)
    #aux.columns = aux.columns.droplevel(0)
    aux = aux.rename_axis(None, axis=1)
    return(aux)

multi_plot(agrup_graf(res_m, geo='REGIAO3'))
